split storage, toast, and utils modules
