+++ src/options/options.html
@@ -123,7 +123,7 @@
     <!-- Toast notifications -->
     <div class="toast-container" id="toastContainer"></div>
 
-    <script src="options.js"></script>
+    <script type="module" src="options.js"></script>
     <script src="../theme-toggle/theme-toggle.js"></script>
   </body>
 </html>
