+++ src/options/utils.js
@@ -1,0 +1,3 @@
+export function normalizeUrl(url) {
+  return url.trim().toLowerCase();
+}
